requirements.txt
Django==2.0.1
gunicorn==19.7.1
Pillow==5.0.0
pytz==2017.3
whitenoise==3.2.1
dj-database-url==0.4.2
psycopg2
django-rest-framework-social-oauth2==1.0.4
